CFS works on Windows again
Build works on Windows with MSVC2019/IF2020 or Intel2020

Summary of changes:

    in the warning regarding OpenMP and MS compilers, switched a "cerr" output to "cout" (so that warning message should be more meaningful)

    deactivated echo output on cfs.bat

    the setting of CFS_NUM_THREADS for MS compilers is handled at a single central place in Environment.cc only.

    DesignElement.cc: adjusted inlining, as __atribute__((always_inline)) is not supported under windows
    StdVector.cc: adjusted limit settings, distinguish between windows and unix/linux
    Matrix.cc: removed some superfluous/erroneous  instantiatons in case of bool matrices

    treated issues that have been created due to restrictions in OpenMP specs of MSVC
    - index in for loops must not be an unsigned int
    - in case of -Werror=sign-compare (gcc7 setting in pipeline) this lead to an error
    - in many cases these only show up in case of c++ and not if g++ is used
    - now casts have been introduced to resolve the compare problem

    fixed a missing URL for zlib

    adjusted file endings for a bunch of patch files (need to have linux format for patching)
    - flann patches have now Unix file format also under windows (required for patch to work)

    boost project has been refined to account for actual Visual Studio version (this is still somewhat messy and needs improvement/clean-up)
    adjusted gedate.bat to make it work for both english and german windows versions (and thus avoid nasty cmake error messages)
    boost-post_install.cmake.in:  fixed a wrong target directory for include files on windows
    .gitattributes: addressed the CRLF/LF problem for windows/linux development environments: due to patching system, some files need LF instead of CRLF even on windows (or else some patches will be rejected....)

+++ source/Domain/Mesh/NcInterfaces/MortarInterface.cc
@@ -419,8 +419,8 @@ void MortarInterface::MoveInterface() {
 
   if ( useGlobalCoords ) {
     #pragma omp parallel for num_threads(CFS_NUM_THREADS)
-    for (UInt i = 0; i < numNodes; ++i) {
-      for (UInt j = 0; j < dim; ++j) {
+    for (Integer i = 0; i < (Integer) numNodes; ++i) {
+      for (Integer j = 0; j < (Integer) dim; ++j) {
         nodeOffsets[i*dim+j] = mParser_->Eval(mphOffset_[j]);
       }
     }
@@ -435,7 +435,7 @@ void MortarInterface::MoveInterface() {
       moveInterfaceOrigCoords_.Resize(numNodes);
       moveInterfaceLocalCoords_.Resize(numNodes);
       #pragma omp parallel for num_threads(CFS_NUM_THREADS)
-      for (UInt i = 0; i < numNodes; ++i) {
+      for (Integer i = 0; i < (Integer) numNodes; ++i) {
         ptGrid_->GetNodeCoordinate(moveInterfaceOrigCoords_[i], nodeNums_[i], false);
         coordSys_->Global2LocalCoord(moveInterfaceLocalCoords_[i], moveInterfaceOrigCoords_[i]);
       }
@@ -443,7 +443,7 @@ void MortarInterface::MoveInterface() {
     }
 
     #pragma omp parallel for private (coordNew,coordTmp) num_threads(CFS_NUM_THREADS)
-    for (UInt i = 0; i < numNodes; ++i) {
+    for (Integer i = 0; i < (Integer) numNodes; ++i) {
       coordNew.Resize(dim);
       coordTmp.Resize(dim);
       for (UInt j = 0; j < dim; ++j) {
@@ -599,7 +599,7 @@ void MortarInterface::UpdateInterface() {
         StdVector< Vector<Double> > p1Rot;
         StdVector< Vector<Double> > p2Rot;
         #pragma omp for
-        for (UInt i = 0; i < intersectionCandiatesIdx_.size(); ++i) {
+        for (Integer i = 0; i < (Integer) intersectionCandiatesIdx_.size(); ++i) {
           UInt mIdx = intersectionCandiatesIdx_[i].first;
           UInt sIdx = intersectionCandiatesIdx_[i].second;
           SurfElem* m_el = masterElems_[mIdx];
@@ -707,7 +707,7 @@ void MortarInterface::PreComputeIntersectionCandidatesCGAL(const StdVector<SurfE
     masterBoxes_.resize(numMasterElems);
     uniqueIdxMaster_.Resize(numMasterElems);
     #pragma omp parallel for num_threads(CFS_NUM_THREADS)
-    for(UInt aBox = 0; aBox < numMasterElems; aBox++){
+    for(Integer aBox = 0; aBox < (Integer) numMasterElems; aBox++){
       boost::array<Double,6> bbox;
       ptGrid_->CreateBBoxFromElement(masterElems[aBox], tolRel_, &bbox[0],isMoving_);
       uniqueIdxMaster_[aBox] = aBox;
@@ -731,7 +731,7 @@ void MortarInterface::PreComputeIntersectionCandidatesCGAL(const StdVector<SurfE
     slaveBoxes_.resize(numSlaveElems);
     uniqueIdxSlave_.Resize(numSlaveElems);
     #pragma omp parallel for num_threads(CFS_NUM_THREADS)
-    for(UInt aBox = 0; aBox < numSlaveElems; aBox++){
+    for(Integer aBox = 0; aBox < (Integer) numSlaveElems; aBox++){
       boost::array<Double,6> bbox;
       uniqueIdxSlave_[aBox] = aBox;
       ptGrid_->CreateBBoxFromElement(slaveElems[aBox], tolRel_, &bbox[0],isMoving_);
